20230323 Before add viewTrace

+++ Models/LabelsModel.cs
@@ -9,7 +9,7 @@ namespace OrderManagementMVC.Models
     {
         public int LabelId { get; set; }
         public string LabelName { get; set; }
-        public double Heigth { get; set; }
-        public double Width { get; set; }
+        public float Heigth { get; set; }
+        public float Width { get; set; }
     }
 }
